Allow setting id on Dialog and Stack. (#171)
* Allow setting id on Dialog and Stack. On Dialog, we have to add the id prop conditionally, because it behaves differently in these two cases:
- Providing id={undefined}
- Not providing id at all

In the latter case, Radix will auto-generate an id (which may be used internally for its aria accessibility props). If providing explicit id={undefined}, then the id is omitted entirely, which may impact accessibility. Therefore the changes in Dialog.tsx look a bit bigger (but all it does is conditionally render the <Content> component with or without id prop based on whether p.id is given)

* Simplify Dialog props

* Clarify name