fixed selection in data table
The code for updating and displaying selection in the data table was
completely faulty. It used index lookup on the data (`data[id]`) to find
the correct row to display/update. This would only work if you
coincidentally used the index in the list as the row key. For other key
values this would fail (error, or possibly select the incorrect row when
you pressed the checkbox).

Changed the code to find the relevant row based on the actual key field.